Baseline-relative features (per-person calibration), ablation, calibrated metrics, service baseline contract
Absolute features gave subject-wise OOF AUC ~0.6; z-scoring every feature against the
subject's own 60-s calibration segment lifts it to ~0.88 (recording AUC ~0.99).
The service now requires a baseline (or baseline_seconds) and reports mean-shift drift
per request instead of a noisy single-session PSI.

+++ tests/test_models_and_validation.py
@@ -14,8 +14,8 @@ def test_subject_cv_never_mixes_subjects(small_dataset):
     # encode the subject id into the "prediction" so we can recover which subjects were
     # in each test fold and check they never appear in the corresponding train fold
     def fp(Xtr, ytr, Xte):
-        tr_ids = {int(round(v)) for v in Xtr[:, 0, 0]}
-        te_ids = {int(round(v)) for v in Xte[:, 0, 0]}
+        tr_ids = {round(v) for v in Xtr[:, 0, 0]}
+        te_ids = {round(v) for v in Xte[:, 0, 0]}
         assert tr_ids.isdisjoint(te_ids)
         return np.full(len(Xte), 0.5)
